Message class, put messages from 1 file in a list
